refactor: move resolveLeaderboard
